+++ click_odoo/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2018 ACSONE SA/NV (<http://acsone.eu>)
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 
-from .cli import env_options  # noqa
+from .cli import CommandWithOdooEnv  # noqa
 from .env import OdooEnvironment  # noqa
 from .env import odoo  # noqa
 from .env import odoo_bin  # noqa